simulator/requirements.txt
# Simple Stream - Event Simulator Requirements
# Author: SE Community
# Purpose: Dependencies for Snowpipe Streaming SDK event simulator
# Expires: 2025-12-24

snowflake-ingest>=2.0.0
cryptography>=41.0.0

